refactor: Removed useless map refactor: Renamed component refactor: Minor fixes
