--- src/Table.css
@@ -1,146 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Rubik:wght@300;500&display=swap");
-:root {
-  --custom-color-table: lightseagreen; }
-
-table.rtr-table {
-  font-family: "Rubik", sans-serif;
-  font-weight: 300;
-  margin: 0;
-  padding: 1rem;
-  box-shadow: 0 -5px 20px rgba(0, 0, 0, 0.1);
-  border: solid 1px #fff;
-  background-color: #fff;
-  border-radius: 1rem;
-  position: relative; }
-  @media screen and (min-width: 600px) {
-    table.rtr-table.custom-color tr:hover td {
-      color: var(--custom-color-table); } }
-  @media screen and (max-width: 600px) {
-    table.rtr-table.custom-color tr:hover td {
-      background-color: var(--custom-color-table); } }
-  @media screen and (max-width: 600px) {
-    table.rtr-table.custom-color td:not(:first-child):before {
-      color: var(--custom-color-table); } }
-  table.rtr-table.dark-mode {
-    filter: invert(0.9); }
-  table.rtr-table.bordered {
-    border: solid 1px #ddd;
-    border-top: none;
-    border-left: none;
-    border-radius: 0;
-    box-shadow: none;
-    padding: 0; }
-    table.rtr-table.bordered th,
-    table.rtr-table.bordered tr,
-    table.rtr-table.bordered td {
-      border: solid 1px #ddd;
-      border-bottom: none;
-      border-right: none;
-      border-radius: 0; }
-    table.rtr-table.bordered tr:nth-child(even) {
-      background-color: #fafafa; }
-    @media screen and (max-width: 600px) {
-      table.rtr-table.bordered {
-        border: none;
-        border-bottom: solid 1px #ddd; }
-        table.rtr-table.bordered th,
-        table.rtr-table.bordered td {
-          border: none; }
-        table.rtr-table.bordered tr {
-          border: none;
-          border-top: solid 1px #ddd;
-          border-bottom: solid 1px #ddd; }
-          table.rtr-table.bordered tr td:first-child {
-            border-top: solid 1px #ddd; } }
-  table.rtr-table thead {
-    text-align: left;
-    width: 100%;
-    position: sticky;
-    top: 0;
-    z-index: 1; }
-    table.rtr-table thead th {
-      padding: 1rem;
-      font-size: 90%;
-      border-bottom: solid 2px #fff;
-      border: solid 1px #fff;
-      background-color: rgba(238, 238, 238, 0.5);
-      backdrop-filter: blur(5px);
-      border-radius: 0.5rem; }
-    @media screen and (max-width: 600px) {
-      table.rtr-table thead tr {
-        background-color: rgba(255, 255, 255, 0.5);
-        backdrop-filter: blur(5px);
-        display: block; }
-      table.rtr-table thead th {
-        display: none; } }
-  table.rtr-table tbody td {
-    border-top: solid 1px #fff;
-    border-bottom: solid 1px #fff;
-    border-radius: 0.5rem;
-    padding: 1rem;
-    text-align: left;
-    word-break: break-word; }
-  table.rtr-table tbody tr:hover td {
-    box-shadow: inset 0 0 0 2px white;
-    background-color: rgba(238, 238, 238, 0.3);
-    color: var(--custom-color-table); }
-  @media screen and (max-width: 600px) {
-    table.rtr-table tbody td {
-      display: block;
-      text-align: left;
-      transition: all 0.2s ease-in-out;
-      position: relative;
-      background: #eee; }
-      table.rtr-table tbody td:before {
-        content: attr(data-label);
-        text-transform: uppercase;
-        font-weight: 700;
-        margin-right: 0.5rem;
-        opacity: 0.5; }
-      table.rtr-table tbody td:after {
-        content: " ";
-        position: absolute;
-        right: 1.2rem;
-        left: auto;
-        top: 1rem;
-        bottom: auto;
-        height: 10px;
-        width: 10px;
-        border-bottom: solid 1px #333;
-        border-right: solid 1px #333;
-        transform: rotate(45deg);
-        pointer-events: none; }
-    table.rtr-table tbody td:not(:first-child) {
-      position: relative;
-      height: 0;
-      padding: 0 1rem;
-      opacity: 0;
-      overflow: hidden; }
-      table.rtr-table tbody td:not(:first-child):before {
-        content: attr(data-label);
-        font-size: 12px;
-        font-weight: 500;
-        color: seagreen;
-        text-transform: uppercase;
-        position: absolute;
-        top: -3px; }
-      table.rtr-table tbody td:not(:first-child):after {
-        content: none; }
-    table.rtr-table tbody tr.open td {
-      height: auto;
-      overflow: visible;
-      opacity: 1;
-      color: white;
-      background-color: var(--custom-color-table); }
-      table.rtr-table tbody tr.open td:first-child {
-        margin-bottom: 0.5rem; }
-      table.rtr-table tbody tr.open td:after {
-        border-bottom: none;
-        border-right: none;
-        border-top: solid 1px white;
-        border-left: solid 1px white;
-        top: 1.4rem; }
-    table.rtr-table tbody tr.open td:not(:first-child) {
-      background-color: white;
-      color: #333;
-      padding: 0.7rem 1rem 0.5rem; } }
